Refactor Account Page
