--- src/pages/Account/AccountContent/index.module.scss
@@ -1,183 +1,0 @@
-@import 'App.scss';
-
-.container {
-  padding: 10rem 0 4rem 0;
-}
-
-.title {
-  margin-bottom: 1rem;
-}
-
-.welcome_wrapper {
-  border-bottom: 0.2rem solid #cfcfcf;
-  display: flex;
-  justify-content: space-between;
-  padding-bottom: 2rem;
-  margin-bottom: 2rem;
-  line-height: normal;
-}
-
-.greeting {
-  font-weight: 700;
-  font-size: 2rem;
-  align-self: flex-end;
-  line-height: 1;
-}
-
-.logout_button {
-  border: 0.2rem solid #000;
-  border-radius: 1rem;
-  text-transform: uppercase;
-  background: #fff;
-  color: #000;
-  padding: 1rem 4rem;
-  font-size: 1.6rem;
-  letter-spacing: $ls-button;
-
-  -webkit-user-select: none;
-  -webkit-touch-callout: none;
-  -moz-user-select: none;
-  -ms-user-select: none;
-  user-select: none;
-
-  -moz-transition: all 0.1s ease-in;
-  -o-transition: all 0.1s ease-in;
-  -webkit-transition: all 0.1s ease-in;
-  transition: all 0.1s ease-in;
-
-  &:hover {
-    cursor: pointer;
-    background-color: $hover-button;
-    color: #fff;
-  }
-
-  &:active {
-    transform: scale(0.9);
-  }
-}
-
-.content_container {
-  display: flex;
-  flex-direction: column;
-  gap: 3.5rem;
-
-  @media (min-width: 768px) {
-    display: grid;
-    grid-template-columns: 2fr 1fr;
-    gap: 4rem;
-  }
-}
-
-.orders_list_wrapper {
-  display: flex;
-  flex-direction: column;
-  gap: 3rem;
-  padding-bottom: 4rem;
-  border-bottom: 0.2rem solid #cfcfcf;
-
-  @media (min-width: 768px) {
-    padding-right: 4rem;
-    border-right: 0.2rem solid #cfcfcf;
-    padding-bottom: 0;
-    border-bottom: none;
-  }
-
-  p {
-    font-weight: 700;
-    font-size: 2.2rem;
-    letter-spacing: $ls-button;
-  }
-}
-
-.sidebar {
-  display: flex;
-  flex-direction: column;
-  gap: 6rem;
-}
-
-.profile_wrapper,
-.addresses_wrapper {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.profile_title,
-.addresses_titles {
-  text-transform: uppercase;
-}
-
-.profile_data {
-  font-family: 'Montserrat', sans-serif;
-  font-size: 1.6rem;
-  color: #3d3d3d;
-
-  li {
-    border: 0.2rem solid #3d3d3d;
-    border-radius: 1rem;
-    padding: 1rem;
-    margin-bottom: 0.5rem;
-    display: flex;
-    align-items: center;
-    gap: 1rem;
-  }
-}
-
-.addresses_list {
-  display: flex;
-  flex-direction: column;
-  gap: 2rem;
-
-  div {
-    padding: 3rem;
-    border: 0.2rem solid #000;
-    border-radius: 1rem;
-    background: #fff;
-    color: #000;
-    font-size: 1.4rem;
-    -moz-transition: all 0.1s ease-in;
-    -o-transition: all 0.1s ease-in;
-    -webkit-transition: all 0.1s ease-in;
-    transition: all 0.1s ease-in;
-  }
-
-  li,
-  h4 {
-    font-family: 'Montserrat', sans-serif;
-    font-size: 1.6rem;
-    color: #3d3d3d;
-  }
-}
-
-.edit_button {
-  padding: 1rem 2rem;
-  margin-top: 1rem;
-  font-size: 1.6rem;
-  border: solid 0.2rem $primary-button;
-  border-radius: 1rem;
-  width: 100%;
-  background: $primary-button;
-  color: #fff;
-  text-transform: uppercase;
-  letter-spacing: $ls-button;
-
-  -moz-transition: all 0.1s ease-in;
-  -o-transition: all 0.1s ease-in;
-  -webkit-transition: all 0.1s ease-in;
-  transition: all 0.1s ease-in;
-
-  &:active {
-    transform: scale(0.9);
-  }
-
-  &:hover {
-    background: $hover-button;
-    color: #fff;
-    border-color: $hover-button;
-  }
-}
-
-a.edit_button {
-  display: flex;
-  justify-content: center;
-}
