+++ src/pages/Account/AccountOrders/index.module.scss
@@ -1,0 +1,22 @@
+@import 'App.scss';
+
+.orders_list_wrapper {
+  display: flex;
+  flex-direction: column;
+  gap: 3rem;
+  padding-bottom: 4rem;
+  border-bottom: 0.2rem solid #cfcfcf;
+
+  @media (min-width: 768px) {
+    padding-right: 4rem;
+    border-right: 0.2rem solid #cfcfcf;
+    padding-bottom: 0;
+    border-bottom: none;
+  }
+
+  p {
+    font-weight: 700;
+    font-size: 2.2rem;
+    letter-spacing: $ls-button;
+  }
+}
