+++ src/pages/Account/OrderModal/index.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const OrderModal = () => {
+  return <div>OrderModal</div>;
+};
+
+export default OrderModal;
