+++ src/pages/Account/index.module.scss
@@ -1,0 +1,75 @@
+@import 'App.scss';
+
+.container {
+  padding: 10rem 0 4rem 0;
+}
+
+.title {
+  margin-bottom: 1rem;
+}
+
+.welcome_wrapper {
+  border-bottom: 0.2rem solid #cfcfcf;
+  display: flex;
+  justify-content: space-between;
+  padding-bottom: 2rem;
+  margin-bottom: 2rem;
+  line-height: normal;
+}
+
+.greeting {
+  font-weight: 700;
+  font-size: 2rem;
+  align-self: flex-end;
+  line-height: 1;
+}
+
+.logout_button {
+  border: 0.2rem solid #000;
+  border-radius: 1rem;
+  text-transform: uppercase;
+  background: #fff;
+  color: #000;
+  padding: 1rem 4rem;
+  font-size: 1.6rem;
+  letter-spacing: $ls-button;
+
+  -webkit-user-select: none;
+  -webkit-touch-callout: none;
+  -moz-user-select: none;
+  -ms-user-select: none;
+  user-select: none;
+
+  -moz-transition: all 0.1s ease-in;
+  -o-transition: all 0.1s ease-in;
+  -webkit-transition: all 0.1s ease-in;
+  transition: all 0.1s ease-in;
+
+  &:hover {
+    cursor: pointer;
+    background-color: $hover-button;
+    color: #fff;
+  }
+
+  &:active {
+    transform: scale(0.9);
+  }
+}
+
+.content_container {
+  display: flex;
+  flex-direction: column;
+  gap: 3.5rem;
+
+  @media (min-width: 768px) {
+    display: grid;
+    grid-template-columns: 2fr 1fr;
+    gap: 4rem;
+  }
+}
+
+.sidebar {
+  display: flex;
+  flex-direction: column;
+  gap: 6rem;
+}
